fix: Pass correct dataType in StructProduct to support UnsafeRow data
StructProduct's eval and doGenCode paths passed null as the dataType
when copying struct fields via InternalRow.get(), which crashed with NPE
for UnsafeRow (Tungsten format from createDataFrame). Fixed by passing
the actual field data types.

Restructured StructProduct tests to run in both codegen and interpreted
modes via an abstract base class, replacing the separate codegen-disabled
test method.
