+++ encoders/src/test/java/au/csiro/pathling/encoders/ExpressionsInterpretedTest.java
@@ -1,0 +1,46 @@
+/*
+ * This is a modified version of the Bunsen library, originally published at
+ * https://github.com/cerner/bunsen.
+ *
+ * Bunsen is copyright 2017 Cerner Innovation, Inc., and is licensed under
+ * the Apache License, version 2.0 (http://www.apache.org/licenses/LICENSE-2.0).
+ *
+ * These modifications are copyright 2018-2026 Commonwealth Scientific
+ * and Industrial Research Organisation (CSIRO) ABN 41 687 119 230.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package au.csiro.pathling.encoders;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+
+/** Tests for FHIR encoder expressions with whole-stage codegen disabled (interpreted eval path). */
+public class ExpressionsInterpretedTest extends ExpressionsBothModesTest {
+
+  /** Set up Spark with whole-stage codegen disabled. */
+  @BeforeAll
+  static void setUpInterpreted() {
+    setUp();
+    spark.conf().set("spark.sql.codegen.wholeStage", "false");
+  }
+
+  /** Restore codegen setting and stop Spark. */
+  @AfterAll
+  static void tearDown() {
+    spark.conf().set("spark.sql.codegen.wholeStage", "true");
+    spark.stop();
+  }
+}
